# openapi.yaml
openapi: 3.0.0
info:
  title: URL Shortener Service
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /url:
    post:
      tags:
        - 'URL'
      summary: Shortens the provided URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  description: The URL to be shortened
                  example: https://www.google.com
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: The shortened URL
                    example: suKX7hF
        '400':
          description: Bad Request
    get:
      tags:
        - 'URL'
      summary: Get all shortened URLs
      parameters:
        - in: query
          name: limit
          required: false
          description: Maximum number of objects to return
          schema:
            type: integer
            minimum: 10
        - in: query
          name: skip
          required: false
          description: Number of objects to skip before starting results
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  urls:
                    type: array
                    items:
                      type: object
                      properties:
                        alias:
                          type: string
                          description: Name of the object
                          example: suKX7hF
                        original:
                          type: string
                          description: Original URL
                          example: https://www.google.com
                        createdAt:
                          type: string
                          description: Created date
                          example: 2024-05-09T06:22:58.424Z

  /url/{alias}:
    get:
      tags:
        - 'URL'
      summary: Gets the original URL and redirects
      parameters:
        - in: path
          name: alias
          schema:
            type: string
          required: true
          description: Alias of a shortened URL
          example: suKX7hF
      responses:
        '307':
          description: Temporary Redirect
          headers:
            Location:
              description: URL to redirect to
              schema:
                type: string
                example: https://www.google.com
        '404':
          description: Not Found